refactor : 구버전 회원 정보 조회 페이지 삭제
