+++ src/pages/auth/Recovery/RecoveryConfirmModal.jsx
@@ -17,7 +17,7 @@ function RecoveryConfirmModal({ onClose, id }) {
     try {
       const response = await axiosInstance.post(`/auth-service/auth/recover/${id}`);
       if (response.data.success) {
-        showSuccess('복구가 완료되었습니다. 다시 로그인 하십시요');
+        showSuccess('복구 되었습니다. 다시 로그인 해주세요.');
         onClose();
         navigate('/');
       } else {
@@ -31,11 +31,10 @@ function RecoveryConfirmModal({ onClose, id }) {
   };
 
   return (
-    <ModalWrapper onClose={onClose} title="계정 복구 확인" size="lg">
+    <ModalWrapper onClose={onClose} title="계정 복구 확인" size="md">
       <div className={styles.RecoveryConfirmModal}>
-        <p className={styles.description}>
-          사용자님의 계정은 휴먼모드입니다. 계정을 복구 하시겠습니까?
-        </p>
+        <p className={styles.description}>사용자님의 계정은 휴먼모드입니다.</p>
+        <p className={styles.description}>계정을 복구 하시겠습니까?</p>
         <div className={styles.buttonGroup}>
           <button type="button" onClick={onClose} disabled={isLoading} className={styles.cancel}>
             돌아가기
