+++ src/pages/auth/Recovery/RecoveryConfirmModal.module.scss
@@ -8,12 +8,13 @@
   justify-content: center;
   align-items: center;
   gap: 30px;
-  margin-top: 70px;
+  margin-top: 30px;
 }
 
 .description {
   font-size: 18px;
   font-weight: bold;
+  margin-bottom: 10px;
 }
 
 .cancel {
