+++ src/pages/user/Info/component/Delete/DeleteConfilmModal.module.scss
@@ -8,18 +8,18 @@
 .description {
   font-size: 25px;
   font-weight: bold;
-  margin-bottom: 10px;
 }
 
 .warning {
   font-size: 14px;
   font-weight: bold;
-  margin-bottom: 80px;
+  margin-top: 20px;
+  margin-bottom: 40px;
 }
 
 .buttonGroup {
   display: flex;
-  gap: 60px;
+  gap: 30px;
 }
 
 .buttonone {
