--- src/pages/user/UserInfo/components/PasswordChangeModal.module.scss
@@ -1,58 +1,0 @@
-.description {
-  color: rgb(0, 0, 0);
-  margin-bottom: 20px;
-  line-height: 1.5;
-  text-align: center;
-  font-weight: bold;
-}
-
-.formselection {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.requirements {
-  margin-bottom: 20px;
-}
-
-.requirementsTitle {
-  font-weight: bold;
-  margin-bottom: 10px;
-  font-size: 10px;
-  margin-left: 5px;
-}
-
-.requirementsList {
-  margin-left: 10px;
-  color: rgb(0, 0, 0);
-  font-size: 10px;
-}
-
-.submitButton {
-  --bg: #000;
-  --hover-bg: #ffffff;
-  --hover-text: #000;
-
-  color: #fff;
-  cursor: pointer;
-  border: 1px solid var(--bg);
-  border-radius: 4px;
-  padding: 0.8em 2em;
-  background: var(--bg);
-  transition: 0.2s;
-  width: 100%;
-
-  &:hover {
-    color: var(--hover-text);
-    transform: translate(-0.25rem, -0.25rem);
-    background: var(--hover-bg);
-    box-shadow: 0.25rem 0.25rem var(--bg);
-  }
-
-  &:disabled {
-    cursor: not-allowed;
-    opacity: 0.7;
-  }
-}
