--- src/pages/user/UserInfo/components/PasswordConfirmModal.module.scss
@@ -1,46 +1,0 @@
-.description {
-  color: rgb(0, 0, 0);
-  margin-bottom: 35px;
-  line-height: 1.5;
-  text-align: center;
-  font-weight: bold;
-}
-
-.formselection {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.submitButton {
-  --bg: #000;
-  --hover-bg: #ffffff;
-  --hover-text: #000;
-
-  color: #fff;
-  cursor: pointer;
-  border: 1px solid var(--bg);
-  border-radius: 4px;
-  padding: 0.8em 2em;
-  background: var(--bg);
-  transition: 0.2s;
-  width: 80%;
-
-  &:hover {
-    color: var(--hover-text);
-    transform: translate(-0.25rem, -0.25rem);
-    background: var(--hover-bg);
-    box-shadow: 0.25rem 0.25rem var(--bg);
-  }
-
-  &:active {
-    transform: translate(0);
-    box-shadow: none;
-  }
-
-  &:disabled {
-    cursor: not-allowed;
-    opacity: 0.7;
-  }
-}
