--- src/pages/user/UserInfo/components/index.js
@@ -1,10 +1,0 @@
-export { default as PasswordInput } from './PasswordInput';
-export { default as EditableField } from './EditableField';
-export { default as ProfileImage } from './ProfileImage';
-export { default as LoadingSpinner } from './LoadingSpinner';
-export { default as ErrorDisplay } from './ErrorDisplay';
-export { default as PhoneVerification } from './PhoneVerification';
-export { default as PasswordConfirmModal } from './PasswordConfirmModal';
-export { default as PasswordChangeModal } from './PasswordChangeModal';
-export { default as WithdrawalModal } from './WithdrawalModal';
-export { default as AddrConfirmModal } from './AddrConfirmModal';
